+++ features/support/test_helper.rb
@@ -77,21 +77,29 @@ module HelperMethods
     end
   end
 
-  def red_alert_box_message
-    page.find(".alert-error").text
+  def p_tag_message
+    page.find('p').text
   end
 
-  def green_alert_box_message
-    page.find(".alert-success").text
+  def red_error_under_textfield
+    page.find(".formError").text
   end
+
+  # def red_alert_box_message
+  #   page.find(".alert-error").text
+  # end
+
+  # def green_alert_box_message
+  #   page.find(".alert-success").text
+  # end
 
   def error_label_css
     ".error"
   end
 
-  def red_error_below_field
-    page.find(".label-error").text
-  end
+  # def red_error_below_field
+  #   page.find(".label-error").text
+  # end
 
   def visit_admin_users_in_new_window
     @admin_tool = AdminTool.new